pyproject.toml
[project]
name = "htp-analyzer-backend"
version = "0.1.0"
description = "HTP Analyzer Backend - YOLOv11 model for HTP (House-Tree-Person) psychological assessment feature detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.15.0",
    "fastapi>=0.116.2",
    "uvicorn>=0.36.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
]
